Updated to v0.3. Now supports AMD. Does a lot more magic.

+++ minionette.js
@@ -1,27 +1,216 @@
-// Minionette.js 0.1.0
+// Minionette.js 0.3
 // (c) 2013 Anthony Bennett/
 // Released under LGPL v3 (see license.txt)
-(function(root, Backbone) {
-	// standard boilerplate stuff
-	var previousMinionette = root.Minionette,
-		Minionette,
-		slice = [].slice,
-		collectionToJSON;
-
-	if ("undefined" == typeof exports) {
-		Minionette = root.Minionette = {};
+(function(root, factory) {
+	// CommonJS; AMD; Register global
+	if ("object" == typeof exports) {
+		factory(exports);
+	} else if (("function" == typeof define) && define.amd) {
+		define(["exports"], factory);
 	} else {
-		Minionette = exports;
+		factory(root.Minionette = {});
 	}
+})(this, function(exports) {
+	// standard boilerplate stuff
+	var Minionette = exports,
+		slice = [].slice;
 
-	Minionette.VERSION = "0.1";
+	Minionette.VERSION = "0.3";
+	Backbone.emulateHTTP = true;
 
-	Minionette.noConflict = function() {
-		root.Minionette = previousMinionette;
-		return this;
+	/////////////////////
+	// MINIONETTE.BINDALL
+	/////////////////////
+	var bindAll = Minionette.bindAll = function(obj) {
+		_.bindAll.apply(obj, [obj].concat(_.functions(obj)));
 	};
 
-	Backbone.emulateHTTP = true;
+	/////////////
+	// DO NOTHING
+	/////////////
+
+	var doNothing = function() {
+	};
+
+	/////////////////////
+	// COLLECTION TO JSON
+	/////////////////////
+
+	var collectionToJSON = function(collection) {
+		// start building result
+		var result = [];
+
+		// loop over collection
+		collection.each(function(item) {
+			result.push(item.toJSON());
+		});
+
+		// done; return result
+		return result;
+	};
+
+	////////////////
+	// SET LISTENERS
+	////////////////
+
+	var setListeners = function(obj, events, context) {
+		_.each(events, function(handler, event) {
+			if (event) {
+				_.each((_.isArray(handler) ? handler : [handler]), function(handler2) {
+					context.listenTo(obj, event, context[handler2]);
+				});
+			}
+		});
+	};
+
+	///////////////////////////
+	// SET APPLICATION BINDINGS
+	///////////////////////////
+
+	var setApplicationBindings = function(events, context) {
+		_.each(events, function(handler, event) {
+			if (event) {
+				_.each((_.isArray(handler) ? handler : [handler]), function(handler2) {
+					var matches = handler2.match(/^model\.(.+)$/);
+					if (matches) {
+						context.application.bind(event, function(value) {
+							context.model.set(matches[1], value);
+						});
+					} else {
+						context.application.bind(event, context[handler2]);
+					}
+				});
+			}
+		});
+	};
+
+	/////////////////
+	// SET COLLECTION
+	/////////////////
+
+	var setCollection = function(collection, events, comparator, context) {
+		// clear previous listeners and reset collection
+		if ("object" == typeof context.collection) {
+			context.stopListening(context.collection);
+			context.collection.reset();
+		}
+
+		// set collection to incoming
+		context.collection = collection;
+
+		// initialize as needed
+		if ("function" == typeof context.collection) {
+			context.collection = new context.collection();
+		}
+
+		// set listeners
+		setListeners(context.collection, events, context);
+
+		// set comparator, if any
+		if (comparator) {
+			this.collection.comparator = comparator;
+			this.collection.sort()
+		}
+	};
+
+	////////////
+	// SET MODEL
+	////////////
+
+	var setModel = function(model, events, context) {
+		// if model is not an instance of a backbone model,
+		// treat it like a plain object and make it so
+		if (("object" == typeof model) && !(model instanceof Backbone.Model)) {
+			model = new Backbone.Model(model);
+		}
+
+		// clear previous listeners
+		// clear previous listeners and reset collection
+		if ("object" == typeof context.model) {
+			context.stopListening(context.model);
+		}
+
+		// set model to incoming
+		context.model = model;
+
+		// set listeners
+		setListeners(context.model, events, context);
+	};
+
+	//////////////
+	// SET VISIBLE
+	//////////////
+
+	var setVisible = function(context) {
+		var type = (typeof context.visible);
+
+		// leave it as-is
+		if ("function" == type) {
+			return;
+		}
+
+		// convert to function if possible
+		if ("string" == type) {
+			var parts = context.visible.split(".");
+			if (parts.length && !_.contains(parts, "") &&
+				("undefined" != typeof context[parts[0]])) {
+				if (2 == parts.length) {
+					// based on a model attribute
+					if ("model" == parts[0]) {
+						context.visible = function() {
+							return context.model.get(parts[1]);
+						};
+						return;
+					}
+					// based on a subproperty / function
+					context.visible = function() {
+						return _.result(context[parts[0]], parts[1]);
+					};
+					return;
+				}
+				// based on a property / function
+				if (1 == parts.length) {
+					context.visible = function() {
+						return _.result(context, parts[0]);
+					};
+					return;
+				}
+			}
+		}
+
+		// otherwise, clear it
+		context.visible = null;
+	};
+
+	///////////////
+	// VIEW MANAGER
+	///////////////
+
+	var ViewManager = function() {
+		this.views = [];
+	};
+	_.extend(ViewManager.prototype, {
+		push: function(view) {
+			this.views.push(view);
+		},
+		remove: function(view) {
+			this.views = _.without(this.views, view);
+			view.remove();
+		},
+		clear: function() {
+			var view;
+			while (view = this.views.pop()) {
+				view.remove();
+			}
+		},
+		invoke: function(method) {
+			_.each(this.views, function(view) {
+				if ("function" == typeof view[method]) {
+					view[method]();
+				}
+			});
+		}
+	});
 
 	/////////////////////////
 	// MINIONETTE.APPLICATION
@@ -31,13 +220,16 @@
 		// keep an internal reference to this;
 		// set up options
 		var application = this,
-			options = _.extend({
+			options = _.defaults((userOptions || {}), {
 				views: application.views,
 				links: application.links
-			}, (userOptions || {}));
+			});
 
 		// make sure this always means this
-		_.bindAll(application);
+		bindAll(application);
+
+		// keep track of views
+		this.views = new ViewManager();
 
 		// initialize each view
 		_.each(options.views, function(view) {
@@ -82,6 +274,10 @@
 
 		// initialize the application itself
 		application.initialize.apply(application, _.values(options));
+
+		// now that everything is initialized run
+		// any after-initialize functions
+		this.views.invoke("afterInitialize");
 	};
 
 	// add an extend method
@@ -89,11 +285,10 @@
 
 	// add other methods
 	_.extend(Minionette.Application.prototype, {
-		views: [],
 		links: [],
 		initialize: function() {},
 		// this is the event bus, which is
-		// borrows heavily from radio.js
+		// borrowed heavily from radio.js
 		events: {},
 		bind: function(name, callback) {
 			if ("undefined" == typeof this.events[name]) {
@@ -134,23 +329,6 @@
 		}
 	});
 
-	/////////////////////
-	// COLLECTION TO JSON
-	/////////////////////
-
-	collectionToJSON = function(collection) {
-		// start building result
-		var result = [];
-
-		// loop over collection
-		collection.each(function(item) {
-			result.push(item.toJSON());
-		});
-
-		// done; return result
-		return result;
-	};
-
 	//////////////////
 	// MINIONETTE VIEW
 	//////////////////
@@ -158,10 +336,13 @@
 	Minionette.View = Backbone.View.extend({
 		constructor: function() {
 			// make sure this always means this
-			_.bindAll(this);
+			bindAll(this);
 
 			// keep a reference to the application
 			this.application = arguments[0].application;
+
+			// bind application events
+			setApplicationBindings((this.applicationEvents || {}), this);
 
 			// call the parent constructor
 			Backbone.View.prototype.constructor.apply(this, arguments);
@@ -174,15 +355,35 @@
 
 	Minionette.ItemView = Backbone.View.extend({
 		constructor: function() {
+			var options = arguments[0];
+
 			// make sure this always means this
-			_.bindAll(this);
+			bindAll(this);
 
 			// keep a reference to the application
-			this.application = arguments[0].application;
+			this.application = options.application;
+
+			// bind application events
+			setApplicationBindings((this.applicationEvents || {}), this);
 
 			// if we have a collection, set it up
 			if (this.collection) {
-				this.setCollection(this.collection);
+				this.setCollection(this.collection, _.defaults(
+					(this.collectionEvents || {}), {
+						"add": "render",
+						"change": "render",
+						"remove": "render",
+						"reset": "render",
+						"sort": "render"
+					}), this.comparator);
+			}
+
+			// compile template
+			if (this.templateEl) {
+				this.template = $(this.templateEl).html();
+			}
+			if (this.template && ("function" != typeof this.template)) {
+				this.template = _.template(this.template);
 			}
 
 			// call parent constructor
@@ -190,61 +391,61 @@
 
 			// if we have a model, set up its events
 			if (this.model) {
-				this.setModel(this.model);
+				this.setModel(this.model, _.defaults(
+					(this.modelEvents || {}),
+					(_.result(this.model.url) ? {
+						"change": "renderAndSave",
+						"destroy": "removeTriggeredByDestroy"
+					} : {
+						"change": "render"
+					})));
 			}
+
+			// set up visible, if any
+			setVisible(this);
 		}
 	});
 
 	// add other methods
 	_.extend(Minionette.ItemView.prototype, {
 		template: null,
-		setCollection: function(collection) {
-			// clear previous events and reset it
-			if ("object" == typeof this.collection) {
-				this.stopListening(this.collection);
-				this.collection.reset();
-			}
-			// set collection to incoming
-			this.collection = collection;
-			// initialize as needed
-			if ("function" == typeof this.collection) {
-				this.collection = new this.collection();
-			}
-
-			// set events
-			this.listenTo(this.collection, "add", this.render);
-			this.listenTo(this.collection, "change", this.render);
-			this.listenTo(this.collection, "remove", this.render);
-			this.listenTo(this.collection, "reset", this.render);
+		templateEl: null,
+		setCollection: function(collection, events, comparator) {
+			setCollection(collection, events, comparator, this);
 		},
-		setModel: function(model) {
-			// clear previous events
-			this.stopListening(this.model);
-			// set model to incoing
-			this.model = model;
-			// set events
-			this.listenTo(this.model, "change", this.renderAndSave);
-			this.listenTo(this.model, "destroy", this.removeTriggeredByDestroy);
+		setModel: function(model, events) {
+			setModel(model, events, this);
 		},
+		beforeRender: function(data) {
+			return data;
+		},
+		afterRender: doNothing,
 		render: function() {
+			// get data
 			var data;
-
-			// compile template
-			if ("function" != typeof this.template) {
-				this.template = _.template($(this.template).html());
-			}
-
-			// run template and append result
 			if (this.collection) {
 				data = { list: collectionToJSON(this.collection) };
 			} else if (this.model) {
-				data = this.model.attributes;
+				data = JSON.parse(JSON.stringify(this.model.attributes));
 			} else {
 				data = null;
 			}
+
+			// run before render function
+			data = this.beforeRender(data);
+
+			// render template if data found
 			if (data) {
 				this.$el.html(this.template(data));
 			}
+
+			// toggle visibility based on value or function
+			if (this.visible) {
+				this.$el.toggle(_.result(this, "visible") ? true : false);
+			}
+
+			// run after render function
+			this.afterRender();
 
 			// allow chaining
 			return this;
@@ -257,13 +458,16 @@
 			});
 			return this;
 		},
+		beforeRemove: doNothing,
 		remove: function(triggeredByDestroy) {
+			this.beforeRemove();
 			this.$el.remove();
 			this.stopListening();
 			if (this.collection) {
 				this.collection = null;
 			} else if (this.model) {
-				if (!triggeredByDestroy) {
+				if (!triggeredByDestroy &&
+					_.result(this.model.url)) {
 					this.model.destroy({
 						success: this.ajaxSuccess,
 						error: this.ajaxError
@@ -307,23 +511,41 @@
 
 	Minionette.ListView = Backbone.View.extend({
 		constructor: function() {
+			var options = arguments[0];
+
 			// make sure this always means this
-			_.bindAll(this);
+			bindAll(this);
 
 			// keep a reference to the application
-			this.application = arguments[0].application;
+			this.application = options.application;
+
+			// bind application events
+			setApplicationBindings((this.applicationEvents || {}), this);
+
+			// keep track of views
+			this.views = new ViewManager();
 
 			// set up collection
-			if ("function" == typeof this.collection) {
-				this.collection = new this.collection();
-			}
+			this.setCollection((this.collection || Backbone.Collection), _.defaults(
+				(this.collectionEvents || {}), {
+					"add": "append",
+					"reset": "render",
+					"sort": "render"
+				}), this.comparator);
 
-			// set up collection events
-			this.listenTo(this.collection, "add", this.append);
-			this.listenTo(this.collection, "reset", this.render);
+			// compile template
+			if (this.templateEl) {
+				this.template = $(this.templateEl).html();
+			}
+			if (this.template && ("function" != typeof this.template)) {
+				this.template = _.template(this.template);
+			}
 
 			// call parent constructor
 			Backbone.View.prototype.constructor.apply(this, arguments);
+
+			// set up visible, if any
+			setVisible(this);
 		}
 	});
 
@@ -331,36 +553,52 @@
 	_.extend(Minionette.ListView.prototype, {
 		itemView: Minionette.ItemView,
 		template: null,
-		elementSelector: null,
+		templateEl: null,
+		itemViewContainer: null,
+		setCollection: function(collection, events, comparator) {
+			setCollection(collection, events, comparator, this);
+		},
 		append: function(item, collection, options) {
 			var view = new this.itemView({
 				application: this.application,
 				model: item
 			});
-			this.$elementSelector.append(view.render().$el);
+			this.views.push(view);
+			(this.$itemViewContainer || this.$el).append(view.render().$el);
 		},
+		beforeRender: function(data) {
+			return data;
+		},
+		afterRender: doNothing,
 		render: function() {
-			var $elementSelector;
+			// run before render function
+			var data = this.beforeRender(this.model);
 
 			// if we have a template...
 			if (this.template) {
-				// compile template
-				if ("function" != typeof this.template) {
-					this.template = _.template($(this.template).html());
-				}
-
 				// run template and append result
-				this.$el.html(this.template(this.model));
-
-				// get item wrapper
-				this.$elementSelector = this.$el.find(this.elementSelector);
-			} else {
-				// item wrapper is just an alias to element
-				this.$elementSelector = this.$el;
+				this.$el.html(this.template(data));
 			}
+
+			// get item wrapper, if any
+			if (this.itemViewContainer) {
+				this.$itemViewContainer = this.$el.find(this.itemViewContainer);
+			}
+
+			// clear out previous views
+			this.views.clear();
 
 			// render collection
 			this.collection.each(this.append);
+
+			// toggle visibility based on value or function
+			if (this.visible) {
+				(this.$itemViewContainer || this.$el).toggle(
+					_.result(this, "visible") ? true : false);
+			}
+
+			// run after render function
+			this.afterRender();
 
 			// allow chaining
 			return this;
@@ -369,28 +607,87 @@
 			return collectionToJSON(this.collection);
 		},
 		load: function(userOptions) {
-			var options = _.extend({
-				reset: true,
-				success: this.ajaxSuccess,
-				error: this.ajaxError
-			}, (userOptions || {}));
-			this.collection.fetch(options);
+			this.collection.fetch(_.defaults(
+				(userOptions || {}), {
+					reset: true,
+					success: this.ajaxSuccess,
+					error: this.ajaxError
+				}));
 		},
 		add: function(attributes, options) {
 			var collection = this.collection,
-				model = new collection.model(attributes, options),
-				success = this.ajaxSuccess,
-				error = this.ajaxError;
+				model = new collection.model(attributes, options);
 
-			model.save(null, {
-				success: function() {
-					collection.add(model);
-					if (success) {
-						success.apply(null, slice.call(arguments));
-					}
-				},
-				error: error
-			});
+			if (collection.url) {
+				var success = this.ajaxSuccess,
+					error = this.ajaxError;
+
+				model.save(null, {
+					success: function() {
+						collection.add(model);
+						if (success) {
+							success.apply(null, slice.call(arguments));
+						}
+					},
+					error: error
+				});
+			} else {
+				collection.add(model);
+			}
+		},
+		setItems: function(items) {
+			this.collection.reset(items);
+		},
+		setComparator: function(comparator) {
+			this.collection.comparator = comparator;
+			this.collection.sort();
 		}
 	});
-}(this, Backbone));
+
+	////////////////////
+	// UNDERSCORE ADDONS
+	////////////////////
+	var moneyRegExp = /^\$?((\d+|\d{1,3}(,\d{3})+)(\.\d{0,2})?|(\d+|\d{1,3}(,\d{3})+)?(\.\d{0,2}))$/;
+	_.money = {
+		validate: function(value) {
+			return ("" + value).match(moneyRegExp);
+		},
+		format: function(amount) {
+			var negative = (amount < 0.0);
+			if (negative) {
+				amount = -amount;
+			}
+
+			// split dollars and cents
+			amount = ("" + amount).split(".");
+
+			// set dollars to 0 if empty
+			if (!amount[0].length) {
+				amount[0] = "0";
+			// and to comma-separated triplets if greater than 3 digits long
+			} else if (amount[0].length > 3) {
+				var amountStart = (amount[0].match(/^(\d{1,3})(\d{3})+$/))[1],
+					amountEnd = amount[0].substring(amountStart.length);
+				amount[0] = [amountStart].concat(amountEnd.match(/\d{3}/g)).join(",");
+			}
+
+			// set cents to 00 if empty
+			if (("undefined" == typeof amount[1]) || !amount[1].length) {
+				amount[1] = "00";
+			// and right pad with zeros if less than 1 digit long
+			} else if (amount[1].length < 2) {
+				amount[1] += "0";
+			// and make sure it doesn't go longer than 2 digits
+			} else if (amount[1].length > 2) {
+				amount[1] = amount[1].substring(0, 2);
+			}
+
+			// prepend $ sign and re-combine dollars and cents
+			return ((negative ? "-$" : "$") + amount.join("."));
+		},
+		unformat: function(money) {
+			// remove extraneous characters and parse as float
+			return parseFloat(("" + money).replace("$", "").replace(/,/g, ""));
+		}
+	};
+});
